used std::atan2

+++ Core/include/Acts/EventData/SpacePointContainer.ipp
@@ -6,8 +6,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <math.h>
-
 namespace Acts {
 
 template <typename container_t, template <typename> class holder_t>
@@ -44,7 +42,7 @@ void SpacePointContainer<container_t, holder_t>::initialize() {
     m_data.setZ(i, external_container.z_impl(i));
     m_data.setRadius(
         i, std::sqrt(m_data.x(i) * m_data.x(i) + m_data.y(i) * m_data.y(i)));
-    m_data.setPhi(i, atan2f(m_data.y(i), m_data.x(i)));
+    m_data.setPhi(i, std::atan2(m_data.y(i), m_data.x(i)));
     m_data.setVarianceR(i, external_container.varianceR_impl(i));
     m_data.setVarianceZ(i, external_container.varianceZ_impl(i));
